+++ src/store/index.js
@@ -43,9 +43,21 @@ export default new Vuex.Store({
         updatingRecipeDirectory: false,
         // Category which is being updated (name)
         categoryUpdating: null,
+        localSettings: {
+            showTagCloudInRecipeList: true,
+        },
     },
 
     mutations: {
+        initializeStore(state) {
+            if (localStorage.getItem("showTagCloudInRecipeList")) {
+                state.localSettings.showTagCloudInRecipeList = JSON.parse(
+                    localStorage.getItem("showTagCloudInRecipeList")
+                )
+            } else {
+                state.localSettings.showTagCloudInRecipeList = true
+            }
+        },
         setAppNavigationRefreshRequired(state, { b }) {
             state.appNavigation.refreshRequired = b
         },
@@ -63,6 +75,10 @@ export default new Vuex.Store({
         },
         setRecipe(state, { r }) {
             const rec = JSON.parse(JSON.stringify(r))
+            if (rec === null) {
+                state.recipe = null
+                return
+            }
             if ("nutrition" in rec && rec.nutrition instanceof Array) {
                 rec.nutrition = {}
             }
@@ -80,6 +96,10 @@ export default new Vuex.Store({
         },
         setSavingRecipe(state, { b }) {
             state.savingRecipe = b
+        },
+        setShowTagCloudInRecipeList(state, { b }) {
+            localStorage.setItem("showTagCloudInRecipeList", JSON.stringify(b))
+            state.localSettings.showTagCloudInRecipeList = b
         },
         setUser(state, { u }) {
             state.user = u
@@ -146,6 +166,9 @@ export default new Vuex.Store({
         },
         setCategoryUpdating(c, { category }) {
             c.commit("setCategoryUpdating", { c: category })
+        },
+        setShowTagCloudInRecipeList(c, { showTagCloud }) {
+            c.commit("setShowTagCloudInRecipeList", { b: showTagCloud })
         },
         updateCategoryName(c, { categoryNames }) {
             const oldName = categoryNames[0]